requirements.txt
Django
django-debug-toolbar
Celery
django-split-settings
django-extensions
django-haystack
djangorestframework
markdown
django-filter
psycopg2
simplejson
Pillow
requests
elasticsearch
pdfminer.six